// PortalAutoservicio/src/app/models/servidores-publicos-eddua.ts
export class servidoresPublicosEddUa{
    CLAVESERVIDORPUBLICO: string;
    NOMBRESERVIDORPUBLICO: string;
    APELLIDOPATERNOSERVIDORPUBLICO: string;
    APELLIDOMATERNOSERVIDORPUBLICO: string;
    RFCSERVIDORPUBLICO: string;
    PUESTOSERVIDORPUBLICO: string;
    TIPOEVALUACION: string;
    ESTATUSEVALUACION: string;
    IDPROCESOVIGENTE: string;
    IDEVALUACIONKP: string;
    CORREOSERVIDORPUBLICO: string;
    IDUNIDAD_ASIGNADA: string
    UNIDAD_ASIGNADA: string
}