// PortalAutoservicio/src/app/models/instrucciones-edd.ts
import { EvaluacionPreguntasRespuesta } from "./detalle-ejecucion-edd-ambos-detalle";
import { ResultadosPreguntas, SeccionesEddPreg } from "./resultados-edd-preg";

export class InstruccionesEDD {
    IDSECCION: string;
    VALOR: string;
    DESCRIPCIONVALOR: string;
    DetalleEjecucionEddambos: ResultadosPreguntas[];
}